fix: removes non-relative imports
Typescript doesn't convert them to relative imports and this breaks
usage in other contexts as browsers and node.
We need babel to do this but also still need typescript in control of
typescript transpilation. The solution is webpack but this will take
quite some effort so using relative imports is the solution for now.